chore: update codebuildId and shift some stuff around
